# docs/swagger.yml
swagger: "2.0"
info:
  description: "カクテルの作り方のAPI"
  version: "1.0.0"
  title: "Cocktails API"
host: "localhost:8080"
tags:
  - name: "cocktails"
    description: "カクテル関連API"
  - name: "shop"
    description: "ショップ関連API"
schemes:
  - "http"
paths:
  /cocktails:
    get:
      tags:
        - "cocktails"
      summary: "カクテルリスト取得API"
      description: "カクテルリストの取得\n offsetとlimitを受け取り、カクテルリストを取得します"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "offset"
          description: "オフセット"
          type: "integer"
          required: false
        - in: "query"
          name: "limit"
          description: "最大取得件数"
          type: "integer"
          required: false
      responses:
        200:
          "description": "A successful response."
          "schema":
            "$ref": "#/definitions/CocktailsListResponse"

    post:
      tags:
        - "cocktails"
      summary: "カクテル登録API"
      description: "カクテルの登録\n カクテル名、材料を登録する"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Request Body"
          required: true
          schema:
            $ref: "#/definitions/CocktailCreateRequest"
      responses:
        200:
          "description": "A successful response."
          "schema":
            "$ref": "#/definitions/CocktailCreateRequest"

  /cocktails/{id}:
    get:
      tags:
        - "cocktails"
      summary: "カクテル情報取得API"
      description: "カクテル情報の取得\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: id
          description: "カクテルID"
          type: integer
          required: true
      responses:
        200:
          "description": "A successful response."
          "schema":
            "$ref": "#/definitions/CocktailResponse"

  /shop:
    post:
      tags:
        - "shop"
      summary: "ショップ登録API"
      description: "ショップの登録\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: "body"
          description: "Request Body"
          required: true
          schema:
            $ref: "#/definitions/ShopRequestBody"
      responses:
        201:
          description: "A successful response."
          schema:
            $ref: "#/definitions/Shop"
    get:
      tags:
        - "shop"
      summary: "ショップ一覧取得API"
      description: "ショップ一覧の取得\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: query
          name: "offset"
          description: "オフセット"
          type: "integer"
          required: false
        - in: query
          name: "limit"
          description: "最大取得件数"
          type: "integer"
          required: false
      responses:
        201:
          description: "A successful response."
          schema:
            $ref: "#/definitions/ShopListResponse"

  /shop/{id}:
    get:
      tags:
        - "shop"
      summary: "ショップ情報取得API"
      description: "ショップ情報の取得\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: id
          description: "ショップID"
          type: integer
          required: true
      responses:
        200:
          description: "A successful response."
          schema:
            $ref: "#/definitions/Shop"

  /shop/{id}/table:
    post:
      tags:
        - "shop"
      summary: "テーブル登録API"
      description: "ショップのテーブル登録\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: id
          description: "ショップID"
          type: integer
          required: true
      responses:
        201:
          description: "A successful response."
          schema:
            $ref: "#/definitions/ShopTable"
    get:
      tags:
        - "shop"
      summary: "テーブル一覧取得API"
      description: "テーブル一覧の取得\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: id
          description: "ショップID"
          type: integer
          required: true
        - in: query
          name: "offset"
          description: "オフセット"
          type: "integer"
          required: false
        - in: query
          name: "limit"
          description: "最大取得件数"
          type: "integer"
          required: false
      responses:
        201:
          description: "A successful response."
          schema:
            $ref: "#/definitions/ShopTableListResponse"

  /shop/{shop_id}/order:
    get:
      tags:
        - "shop"
      summary: "ショップの注文情報取得API"
      description: "ショップの注文情報取得"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: shop_id
          description: "ショップID"
          type: integer
          required: true
        - in: query
          name: unprovided
          description: "未提供の注文かどうか\n trueの場合、未提供の注文情報のみ取得"
          type: boolean
          required: false
        - in: query
          name: offset
          description: "オフセット"
          type: integer
          required: false
        - in: query
          name: limit
          description: "最大取得件数"
          type: integer
          required: false
      responses:
        200:
          description: "A successful response."
          schema:
            $ref: "#/definitions/CocktailList"

  /shop/{shop_id}/table/{table_id}:
    get:
      tags:
        - "shop"
      summary: "テーブル情報取得API"
      description: "テーブル情報の取得\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: shop_id
          description: "ショップID"
          type: integer
          required: true
        - in: path
          name: table_id
          description: "テーブルID"
          type: integer
          required: true
      responses:
        200:
          description: "A successful response."
          schema:
            $ref: "#/definitions/ShopTable"

  /shop/{shop_id}/table/{table_id}/order:
    post:
      tags:
        - "shop"
      summary: "注文API"
      description: "注文\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: shop_id
          description: "ショップID"
          type: integer
          required: true
        - in: path
          name: table_id
          description: "テーブルID"
          type: integer
          required: true
        - in: body
          name: "body"
          description: "Request Body"
          required: true
          schema:
            $ref: "#/definitions/ShopOrderRequest"
      responses:
        201:
          description: "A successful response."
          schema:
            $ref: "#/definitions/ShopOrder"
    get:
      tags:
        - "shop"
      summary: "注文情報取得API"
      description: "注文情報の取得\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: shop_id
          description: "ショップID"
          type: integer
          required: true
        - in: path
          name: table_id
          description: "テーブルID"
          type: integer
          required: true
        - in: query
          name: unprovided
          description: "未提供の注文を取得するフラグ"
          type: boolean
          required: false
      responses:
        200:
          description: "A successful response."
          schema:
            $ref: "#/definitions/CocktailList"

  /shop/{shop_id}/table/{table_id}/order/{order_id}:
    put:
      tags:
        - "shop"
      summary: "注文提供API"
      description: "注文を提供済みにする\n"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: path
          name: shop_id
          description: "ショップID"
          type: integer
          required: true
        - in: path
          name: table_id
          description: "テーブルID"
          type: integer
          required: true
        - in: path
          name: order_id
          description: "注文ID"
          type: integer
          required: true
      responses:
        200:
          description: "A successful response."

definitions:
  Cocktail:
    type: object
    properties:
      name:
        type: string
        description: "カクテル名"
      image_url:
        type: string
        description: "画像URL"
  CocktailList:
    type: object
    properties:
      cocktails:
        type: array
        items:
          $ref: "#/definitions/Cocktail"
  CocktailsListResponse:
    type: "object"
    properties:
      cocktails:
        type: "array"
        items:
          $ref: "#/definitions/CocktailResponse"
  CocktailResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "カクテルID"
      name:
        type: "string"
        description: "カクテル名"
      image_url:
        type: "string"
        description: "画像URL"
      materials:
        type: "array"
        items:
          $ref: "#/definitions/CocktailMaterial"
  CocktailMaterial:
    type: "object"
    properties:
      id:
        type: "string"
        description: "材料ID"
      name:
        type: "string"
        description: "材料名"
      quantity:
        type: object
        $ref: "#/definitions/MaterialQuantity"
  MaterialQuantity:
    type: "object"
    properties:
      quantity:
        type: "integer"
        description: "分量"
      unit:
        type: "string"
        description: "単位"
  CocktailCreateRequest:
    type: "object"
    properties:
      name:
        type: string
        description: "カクテル名"
      image:
        type: object
        description: "カクテル画像"
      materials:
        type: array
        items:
          $ref: "#/definitions/CocktailCreateMaterial"
        description: "材料リスト"
  CocktailCreateMaterial:
    type: object
    properties:
      name:
        type: string
        description: "材料名"
      quantity:
        type: object
        $ref: "#/definitions/MaterialQuantity"

  Shop:
    type: object
    properties:
      id:
        type: integer
        description: "ショップID"
      name:
        type: string
        description: "ショップ名"
  ShopListResponse:
    type: array
    items:
      $ref: "#/definitions/Shop"
  ShopRequestBody:
    type: object
    properties:
      name:
        type: string
        description: "ショップ名"
  ShopTable:
    type: object
    properties:
      id:
        type: integer
        description: "テーブルID"
      shop_id:
        type: integer
        description: "ショップID"
  ShopTableListResponse:
    type: array
    items:
      $ref: "#/definitions/ShopTable"
  ShopOrder:
    type: object
    properties:
      id:
        type: integer
        description: "注文ID"
      table_id:
        type: integer
        description: "テーブルID"
      cocktail:
        type: object
        $ref: "#/definitions/CocktailResponse"
  ShopOrderRequest:
    type: object
    properties:
      cocktail_ids:
        type: array
        description: "カクテルIDリスト"
        items:
          properties:
            cocktail_id:
              type: integer
              description: "カクテルID"